pyproject.toml
[project]
name = "amazonstockprediction"
version = "0.1.0"
description = "Daily Amazon Stock Predictions"
authors = [
    {name = "amulyaprasanth",email = "amulyaprasanth301@gmail.com"}
]
license = {text = "mit"}
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "numpy (<2.0.0)",
    "pandas (<2.2.0)",
    "hopsworks[python] (>=4.1.8,<5.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "ta (>=0.11.0,<0.12.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "dataframe-image (>=0.2.7,<0.3.0)",
    "yfinance (>=0.2.54,<0.3.0)",
    "xgboost (>=2.1.4,<3.0.0)",
    "streamlit (>=1.43.1,<2.0.0)",
    "plotly (>=6.0.0,<7.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"



[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
ipykernel = "^6.29.5"

[tool.poetry]
packages = [
    {include = "amazonstockprediction"}
]
